Refactor nested model
